core/memory_functions:
Implement Movemem, a procedure version of Memmove, which has a C-style semantics.

Since Bcopy is a subprogram with the same functionalities, and is not used
throughout the whole codebase, it is then deleted.

Cpymem used to call Memmove as well, and so it ends up as a duplicate of
the new Movemem. Thus, Cpymem gets modified and now calls Memcpy.
This way the subprogram set looks more homogeneous.

Remember that [mem]move[mem] is safe and works with overlapping memory areas,
[mem]cpy[mem] isn't.