core/memory_functions: re-organize by using GCC's __bultin-memXXX
NOTE: this should be transparent from the user's point of view.

We use a 2-level implementation for memory functions.
Like before, we have visible "C-style" MemXXX function subprograms and "Ada"
XXXmem procedures (furthermore, they are now inlined, unlike before).
Now we use GCC's __builtin-memXXX in their bodies, in order to gain a level of
optimization -- but without the possibility of override them, because at a first
glance it seems to make little sense, due to the hope of GCC emitting optimzed code.
But since GCC does not always optimize these routines, it could ends up to implicitly
call C "standard" library functions like the plain memcpy, memcmp, and so on.
So we move the function implementations towards "E" separable subprograms (only visibile
in the body), which are exported and get called as standard memcpy/memcmp/etc, and can
be this time overridable like the old "legacy" subprograms. In the end, we add a layer
of a (possible) first optimization by using GCC inline builtins.