configuration.in/Makefile/basedefs.gpr/libutils/createconfigure.[sh|gpr]: allow gprbuild to stay anywhere in the filesystem
Since gprbuild is a package with no relationship with the toolchain,
we use either its prefix (i.e., the containing installation directory
of the bin|lib|libexec|share|... hierarchies) or a relative path towards an
undefinite system PATH specification.

This way, you can install/use GPRbuild under any prefix.

Either:

a)
specify GPRBUILD_PREFIX like the windos example in configuration.in

c)
leave GPRBUILD_PREFIX empty in configuration.in, and just make the
<prefix>/bin directory (as usual with every other executable)
of gprbuild visible by setting the PATH variable

c)
leave GPRBUILD_PREFIX empty in configuration.in, then, if you have an
already system gprbuild working, the OS will pick the first available
gprbuild executable and then it will be able to compute the path towards
gprlib using the default relative path

d)
overlay the installation with that of the toolchain

Of course GPRBUILD_PREFIX is irrelevant in GNATMAKE build mode.