RISC-V: restructure some parts
1) more precise definitions of registers, based on the notion of MXLEN

2) mtime/mtimecmp are now in the MTIME unit, with per-implementation
informations.

RISC-V allows an implementation to have a memory-mapped set of registers,
but their address are not preassigned. We import, for now the address of
the registers by specifiyng them in the configuration, by means of linker
symbols. This is not spectacular, but it's also true that those addresses
are dependent on the platform and so on the hardware specs (like an FPGA
implementation of the RISC-V core).

Maybe in the future we will describe per-implementation informations in
configurable subdirectories of the CPU hierarchy (like other families of
CPU, e.g., M68k, MIPS, x86, etc).