NEORV32: new platform
From https://github.com/stnolting/neorv32:

"The NEORV32 Processor is a customizable microcontroller-like system
on chip (SoC) built around the NEORV32 RISC-V CPU and written in
platform-independent VHDL."

This is a quick-and-dirty setup for running SweetAda on a (simulated)
NEORV32 chip, by means of GHDL simulator. If you are brave enough, it
could run on a real FPGA hardware.

1)
Download NEORV32, setup the toolchain as per documentation (sensible
variables are in the sw/common/common.mak file), have GHDL online, and
try to simulate the "hello, World" example program (sw/example/hello_world).
2)
If everything goes fine, build SweetAda, setup the configuration.in
and do a "make all". Then the "make run" will try to execute the code
inside the emulated RISC-V SoC.

The RUN_COMMAND substitutes the hello_world's ELF executable with
that of SweetAda; doing a "make ... sim" should rebuild only
the VHDL image that is going to be emulated; at this point you should
see the welcome message of SweetAda in the console after a few seconds.

A more engineered procedure will be hopefully provided in the future.