LEON3 (QEMU): try to follow both QEMU and the documentation
According to the documentation (GR712RC), the timer interrupt should
be cleared by writing a 1 to the clear register IC array.

That being said, QEMU seems to clear the GPTIMER.CR1.IP pending bit
(in the statement above, that is not changed) by writing a 1,
when the manual states with no doubt that it should be cleared by
writing a 0.

What it's strange is that wrong values do both work in both registers.

We flagged the situation to QEMU maintainers a long time ago by sending
a patch, but with no response. Very annoying.

We don't have the hardware to test, so we stick to this strange
configuration for the sake of running QEMU, until someone will clarify
the situation.