Makefiles/GPRbuild/libutils/createconfigure/RTS/sfp: introduce GNATBIND_SECSTACK
We introduce a more fine-grained control over the SFP's secondary stack memory size.
Firstly, we define a new variable GNATBIND_SECSTACK, definable per-target in
the usual configuration.in. If you define it -- by writing the memory size expressed
as a numer followed by "k" or "M", like in

GNATBIND_SECSTACK := 16k
GNATBIND_SECSTACK := 16M

then, this amount will be assumed as the secondary stack size and will be passed
with the "-D" switch to GNATBIND (before a 4k stack was specified in sweetada.gpr,
regardless of the target).

Otherwise, the memory size will be taken as a default defined in the spec of
the RTS, that is in rts/src/sfp/Makefile.in

The common, current, shared default value (i.e. for all targets) is fixed at 4k
(before was 10k, but the sweetada.gpr's explicit -D switch make the value
practically useless). You can find this value stated also in System.Parametes.ads
as a system-wide setting.

But, in the case of the AVR (and maybe we will extend this also to other
resource-critical targets), the value is overriden, assigning only 128 bytes
in the RTS target own configuration (that is, rts/src/targets/AVR/Makefile.in).

So, if you want control over the secstack memory size:

1)
declare explicitly GNATBIND_SECSTACK in configuration.in (preferred)

2)
declare RUNTIME_DEFAULT_SEC_STACK_SIZE in rts/src/targets/<target>/Makefile.in
and rebuild the RTS. But I tend to see this as a default project-wide property.
Maybe you can suggest a ballpark figure -- good for standard uses.

This way now there is a complete independent target setup.