Makefiles in clibrary/core/cpus/platforms:
Thanks to the symlink synchronization feature in libutils/createsymlink.ps1,
we can avoid to physically copy (in Win machines) the gnat.adc file, instead
of linking it to the source file in the root directory like we do in *ix
machines.

Furthermore, this will let us to simplify a lot all the Makefiles, by using
a common idiom in the configure rule.