RTS/Makefiles:
Thanks to a new feature of the processcfg script, we can perform a
series of simplifications and/or optimizations that also hopefully
increase modularity and ease of use.

1)
We do not call dos2unix anymore, since we perform a textual processing
directly inside the processcfg utility (with a "-r" switch). This is
necessary in order to with' units coming from textual substitution in
Win environments when building with -gnatyd (no DOS LT), e.g., all the
system.ads files.

2)
Due to the 1), we can get rid of dos2unix completely, since it is used
only in the RTS build phase -- anyway -- we keep the utility because it
could be useful for other purposes; but now it's totally unused.

3)
Not calling dos2unix leads to a simplification of rules and recipes in
the Makefile.in fragments contained in the various files located in the
RTS source tree. They end up in manageable fragments, and we want to
use the same logic for common files introduced some commits ago, because
potentially we have to manage not only system.ads but also other units
(e.g., interfac.ads in the case of AVR).

4)
It seems that the same logic of installation for common files that was
implemented in the RTS main Makefile could be used also for the target
files; so move all target actions to a rule that resembles the common one,
also thanks to the segregation of variables introduced some commits ago.

5)
The Makefile.in code fragments in src/sfp, src/zfp and src/targets/<cpu>,
once cleaned up, end up almost in merely "configuration" files, with just
variable assignments; so rename them, from "Makefile.in" to "configuration.in",
because this emphasizes their true nature. But, some rules could be present
in these files, so segregate the pure makefile rules to a "true" Makefile.in.
This case is only present in src/sfp/Makefile.in because the compilation of
a-except needs a special treatment.

6)
Note that rts.cs.in still needs to be a separate file, because it must be
included by the main sweetada toolchain build circuitry, as it may contains
global-system-wide compiler switches that must be used used when compiling
every other unit.