CPUs/cpu.gpr|libcpu.ads: remove libcpu.ads fake spec
The following commits:

- f855ee9c ("Makefile.lb.in, all CPUs: - remove creation of empty library - empty stub .S to silence GPRbuild unhappy about null library - delete GPRbuild dependency on OBJECTS variable", 2023-04-06)
- 0d0d0803 ("libcpu: change previous commit policy about cpu.S due to GPR, use an Ada empty spec", 2023-04-06)

introduced the CPU/*/libcpu.ads fake spec, to make GPRbuild happy when building
the CPU library, since there is no way to build an empty one. Other libraries,
like libcore, libplatform and libclibrary, are non-empty by definition.

This is, because some CPUs -- so far -- don't need implicit code, thus ending up
in an empty libcpu.a. The commits remove the previous Makefile fragment which
builds an empty library, by not calling GPRbuild if no object files are specified
in the Makefile. GNATMAKE build mode does not suffers from this behaviour.

But this is far from ideal, because the results of GNATMAKE and GPRbuild builds
are now different (you have a totally useless libcpu package hanging in some
.ali/.o/.a files). Worse, even if that fact could be mitigated by optimizations
and/or section discarding, almost every target will show different memory
footprints in the first instance. This causes a sense of dissatisfaction.

Before going to explain all the details, please note that this has to do
only with special code needed by the runtime and/or libgcc, not the "normal"
CPU code. Average users should not have problems with that.

So, we reinstate the practice of creating an empty .a library, for the sake
of correctness. But perhaps there is a price to pay (albeit under very special,
non-existent-so-far conditions). Indeed the GPRbuild is called if there are
objects specified in the Makefile. This is due to how SweetAda build system
is conceived. So, if one wants to add files to the CPU library, then, not only
the .gpr should be updated, even the Makefile should be modified, because
they must be kept in-synch.

This is ugly, because it's a redundant procedure, even if it will be as simple
as writing, e.g., "OBJECTS := x" (with an appropriate ifeq() handling logic).

But, together with the previously explained considerations:

1) Makefiles should not be touched, since there are enough configuration.in
   files in order to cover every aspect of the configuration (apart those
   containing parameters, like gnatmake.*.in), freeing the user from complicated
   things like working with Makefile targets

2) libcpu "implicit" handling should be stable, changed only for system-wide
   deep implementation code (libgcc updates, RTS low-level code, etc)

So, in the end, if the casual user wants to do that, we resort to "force" the
build of libcpu in GPRbuild mode (i.e., you have only to add files to "Source_Files"
list of the cpu.gpr), provided an "export GPRBUILD_ENABLE_cpu := Y" is present
in the platform's configuration.in.

This way, the Makefile will call GPRbuild even if there the list of (implicit)
object files in the CPU Makefile is empty.