libutils/gprdeps[bat|ps1|sh]: introduce gprdeps utility
This utility is a very crude parser to detect dependencies in a GPRbuild
project file. It recursively parses with'ing clauses of a root .gpr, printing
on its output a list of files.
With'ing clauses must be listed once per line, should work with both u/l case.
Everything after a non-with clause closes the current parsed file (except
comments and empty lines, of course).

It will be used as an aid in the implementation of GPRbuild configuration
file management.