All Makefiles, Makefiles.tc.in, gnatmake.tc.in: revision of the build system
We integrate the feature of configuration files dependency handling.

Now, whenever a configuration file is changed, a "make clean"/"make configure"
are automatically triggered.

This way, a change of configuration is less error-prone. As a consequence of
this, you could even avoid the "make configure" phase, and you can run directly
a "make all" (but you do not get the infodump). "make configure" is not deprecated,
though, and does not harm.

Configuration files are:
- */configuration.in
- gnatmake_*
- gpubuild_*

The configure.[ads|h] files in every platform are automatically recognized
and processed, provided that you list them in the platform's configuration.in
like:

CONFIGURE_FILES_PLATFORM := configure.ads configure.h

(the *.in template must be existent). Note that those files are NOT configuration
files and do not get tracked by the build system; they are templates, getting
their parameters from configuration.

Also Makefile, Makefile.tc.in and others, are not tracked because they should not
be touched on a daily basis. If you touch them, run explicitly a "make configure"
to clean the build and restart a new cycle with new compiler switches or whatever
you've changed.

Besides, the commit includes a remove of redundant variables, optimizations and
many simplifications for RTS and other stuffes.