Makefile: add preprocessing capabilities to linker script
This way we can preprocess a source linker script (variable LD_SCRIPT)
which must have a ".S" suffix added -- to the filename you've eventually
defined. The only prerequisite is the final linker script ending in
a standard ".lds"

If, e.g., you stick with the SweetAda default one, then a "linker.lds.S"
can be preprocessed (if it exists), else, if you write LD_SCRIPT := <>.lds
in the platform's configuration.in, the filename $(LD_SCRIPT).S (i.e. <>.lds.S)
will be preprocessed (if it exists). If <>.lds.S does not exist, then no
preprocessing will be taken into account and <>.lds should be already in place
as usual.

To make things precise (and a little bit pedantic), you can add the generated
linker.lds to the list of objects to be cleaned up (with a "make clean" action)
by stating:

CLEAN_OBJECTS_COMMON += $(LD_SCRIPT)

(obviously not the ".S" one).