Makefile[.lb.in], configuration.in, platform.gpr, libutils/createconfigure.[ps1|sh]
We introduce some changes and a level of configuration when dealing
with non-Ada source files.

First of all, we decorate the configuration files with comments and a textual separation
of variables in classes, so to have a clean view of assignments. Remember that
the configuration files are user responsibility, the current layout is only an example
and works out practically in all cases, but you are free to rearrange to suit
your needs.

That being said, we rename the variable STARTUP_OBJECTS to LOWLEVEL_FILES_PLATFORM,
which carries a list of non-Ada low-level files. STARTUP_OBJECTS is a bit misnomer,
since it contains source files and not objects in the very sense of the term, and
some files are not only useful for machine startup, like the standard llkernel.S.
In the same way, we rename GCC_SWITCHES_STARTUP to GCC_SWITCHES_LOWLEVEL_PLATFORM.

That being said, the integration of this files is now natural and consistent,
when before was rather cumbersome. As a consequence of this, we wipe out the
variable ADDITIONAL_ASM_OBJECTS (where used) and change the behaviour of
EXTERNAL_OBJECTS. Now EXTERNAL_OBJECTS carry a list of true binary objects
(i.e., a Fortran *.a library) to be linked in the final executable. Just specify
the fully qualified path, then manage the calls to functions in that library
in the usual way (import, etc).

That being said, non-Ada files could be C files as well, and the dependencies
are tracked correctly.

That being said, the handling of all those features should work also in GPRbuild
mode, since we introduce also the detection of source files when we create the
configure.gpr project file. Albeit GPRbuild theoretically prohibites header files
with a shared extension in Asm_Cpp and C files, in the end it seems to be able to
handle this. GNATMAKE mode should work as usual.

The immediate benefit is a clean configuration scheme, where the Makefiles and
the platform.gpr files are nearly identical in all platforms and should not need
user intervention. Anyway we mantain a copy in every platform so we can cover
very special corner cases.