ArduinoUNO: allow simavr emulation
You can run the platform under the simavr emulator, by setting SIMAVR := Y.

To this end, download simavr and set the appropriate installation prefix in the
configuration.in file (SIMAVR_PREFIX). Be sure to download a version that allows
the generation of .vcd dump files, so you can trace ioports like the example
provided.

Otherwise, build yourself simavr from source, enabling in simavr/Makefile:

CFLAGS += -DCONFIG_SIMAVR_TRACE=1

The example will dump the output of PORTB5, which is toggled by the code
in application.adb. Once you have run the emulation, you can do:

$ gtkwave simavr.vcd

and look at the scope trace of the signal.

The output of the UART message is augmented with an LF to let simavr show
a full line without buffering.

Note that the parameters of the vcd dump file are stated in the simavr.c
startup file, which is configured for a distribution version of simavr, if
you use a own-build version, change the include path accordingly. The C
fragment also states a 16 MHz clock to the AVR chip.