libutils/gccdefines: introducing "gccdefines" utility
gccdefines is a utility that runs during configuration, processes GCC output
and create a "GCC_Define" package which contains various macro definitions,
extracting them from the current GCC configuration (which depends on user's
settings).

Think of it as a sort of preprocessing mechanism.

This way, the compiler can optimize section of code if the appropriate flag
is turned on, or off, or some macro exists. If the optimizer is smart enough,
it can entirely wipe out the section of code which is not activated by the macro,
since it is a compile-time constant.

GCC macros get being translated in proper Ada syntax, and rules can be specified
when the package is created. See the utility header for an explanation.

The core complex configuration uses now this mechanism and create as default
the package with CPU definitions and some constants, which will be used in the
next commit for the RISC-V CPU to detect the value of the "atomic" builtins
properties.