Makefile: precise control of linked units
The SweetAda build machinery has the concepts of implicit units, i.e.
units that must be present in the final object even if they are not
with'ed by Ada code, since they are
- implicitly called by the expander (new, exceptions, etc)
- provide LibGCC-like functions (basic arithmetic, etc)

But if they are used explicitly by Ada code, the linker gets two copies
(only from a name-reference point of view) because they are present both
as explicit units, and in the core/cpu libraries.

This is no problem at all since the unit is compiled only once, and only
one reference name will be picked up by the linker.

But, for extremely precision, we delete from the list of explicit units
those that are already present in the implicit library.