Makefile[.ut.in]+libutils/gprbindfilt.[bat|ps1|sh]:
We introduce the gprbindfilt utility.

When executing the bind phase, we redirect the output of GNATBIND in order
to have the availability of elaboration informations in the gnatbind_elab.lst
file. No problems in GNATMAKE mode, in GPRbuild mode there are, just for
a change. Unfortunately we are redirecting also notifications from the binder,
which are intermixed with binder process data, and those notifications end
up in the file, "corrupting" it.

Not really a problem, since these are very negligible details, but we want
absolutely correct behaviour and clean information report in every phase,
because:

a) chances are that maybe someone would parse the file for an automated
   analysis
b) it is ugly anyway

So this script demultiplexes the output of GPRbuild, parsing the informations
about the elaboration order. Furthermore, it filters out the notifications to
stdout, so that we have now a perfect screen information layout, with a
completely transparent behaviour.