Core/CLibrary/others: get rid of Interfaces.C
We have made use of System.Interfaces.C in various places, most of them to
handle low-level interfacing towards the compiler and communication with
C-style objects. This unit defines the basic C types needed to satisfy
convention calls, and the Bits unit (which is with'ed practically everywhere)
imports it.

But with'ing Interfaces.C is evil, because you are forced to drag in a lot
of unnecessary things, which break the elaboration flow.

So, we do the other way round.

Instead of Bits subtyping Bytesize from i-c's size_t, we define a native
Bytesize type which exactly resembles C's size_t type. Then we create Bits.C
which defines the bare minimum set of C types needed to interface low-level
stuff with the compiler.

Once we did that, we have a set of basic low-level units which do not require
a drag of unnecessary objects, and we can make them pragma No_Elaboration_Code.

Obviously Interfaces.C is still present in the codebase, but its use will
be restricted to very particularly cases of user code.