toolchains/GNATTOOLS.sh: new script for GNATTOOLS
This script will build only the GNATMAKE executable for a target toolchain,
with a configurable ALI versions string. It us useful for toolchains that
do not provide the GNATMAKE tool.

See the updated MANUAL.TXT.

Due to this script, we remove the big zip files in the libutils/exe directory,
which are anyway deprecated.