Makefile[tc.in]: changes and optimizations
We remove the concept of "unprefixed" GNATMAKE, since we also provide a
script to build it from source.

In order to reduce build time, various shell fragments are now under
conditionals; this way we detect versions only when needed, and we avoid to
call things like cygpath.exe which are pretty heavy in windos environment.

Optimzied some variable processing and redundant checks.