QEMU-RISC-V:
We can start playing dangerous games with multicore SMP.

The AP (application processor) harts will be triggered by the main hart.
They get a stack and a start procedure from the main hart, printing their
respective hart id on the console.

There are no locking and/or semaphores in SweetAda so far for the RISC-V,
but for now we want only introduce some primitive form of multithreading.

For those who feel skeptical about this approach, the mechanism is used
also in the Linux kernel.