core/[configuration.in|core.gpr]:
In GPRbuild mode we need to communicate to the build machine the CPU files
used when compiling units like mutex, which use CPU-dependent bits of code.

Note that, as usual, all this machinery is needed to not see GPRbuild
keeping to recompile units when it should not.

GNATMAKE mode does not suffer this behavior.