libutils: Windos symlink handling
We introduce the new script rmsymlink, only available in Win platforms,
changing some semantics of createsymlink.

Rationale

SweetAda does make use of symbolic links, in order to gain modularity and
configurability. As an example, the various subplatforms of the PC-x86
platform are files that get symlinked in the platform directory. Also the
application.adb/ads files are symlinked that way.

There are no problems in un*x environments, because the semantics of symbolic
links is well defined. In the Windos environment unfortunately this cannot be
achieved. Symbolic links can be created, but they do not get correct timestamp
behaviour. If you edit a symlinked file, as a matter of fact you edit the
target file, but the symlink itself does not get updated.

This has serious consequences, because the build machinery cannot detect
an updated file, and refuses to compile the file. Even with the appropriate
option supplied to GNU Make, this does not happen, because readlink() and
lstat() system calls are not available in Windos.

So we change the semantics: we do not symlink the files, we physically copy
them (roughly the same thing that happens in Cygwin/MSYS environments).

Now the problem is, if you re-configure from scratch your system (or you
issue a "make distclean" command), then the edited file goes away, because
that type of action will clean up every installed (virtually symlinked, really
copied) file. So your changes are lost.

So we introduce the rmsymlink utility. When a distclean is being issued,
the utility compares the timestamp of the "original" file with that of the
"destination" installed file. If the installed file is more recent, a
prompt will ask if you want to back-update the "origin" file, thus preserving
your work.

Note:
We will slowly insert this "feature" in CPU, platform and application
hierarchies not at once, but where appropriate, one after another until
completion.