share/openocd.[py|tcl]:
We enhance a little bit the OpenOCD interface scripts.

Now we have the switch "-commandfile <filename>", which sources commands
into the OpenOCD interpreter from a file. This way you can specify a very
complicated series of setups before running/debug the target (in the same
way the configuration scripts which comes with OpenOCD are doing). The old
"-command" switch continue to exist and will be executed after this file
sourcing, if it's not empty.

Then, the detected start address of the executable is now made a Tcl variable,
so you can get its value by referring to the "$start_address" name.

Various OpenOCD target platforms will slowly use this feature, avoiding
encoding (possibly long) hardwired setups in configuration.in, keeping
them separate and more manageable. You have not to touch configuration.in
anymore, which is ugly, tedious and causes a recompilation from scratch.

Note: in the OpenOCD command file you can output string and values, please
pay attention that the output is visible in the OpenOCD server window,
not the terminal shell from which you are issue usual make commands.