share/openocd.[py|tcl]: further enhance the interface scripts
We further enhance the driving scripts for OpenOCD operations.

We introduce the "-noload" and "-noexec" switches, which respectively
prevent the script from downloading the executable into the target and
run the CPU.

These enhancements are transparent, and old configurations should work
the same. But now we have a more fine-grained handling to cover special
situations, writing custom initializations in the file specified with the
"-commandfile" switch.

By specifying "-noload", you may want to download for yourself the executable
into the target by means of the standard OpenOCD command "load_image" (see
openocd.cmd examples for, e.g., FRDM-KL46Z/HiFive1/MSP432P401R/STM32F769)
This switch also set the $noload_flag variable (!= 0) in the script, so you
can check if you pass this switch to it.

By specifying "-noexec", you should then issue the OpenOCD "resume" command
(and you know both the executable and the start address because they are in,
respectively, the $sweetada_elf and $start_address variables). This switch
sets the $noexec_flag variable in the same way as above.

The "-debug" flag automatically implies "-noexec" (and, again, sets the
$debug_mode variable), so you can simply issue this switch in the debug
command string inside the configuration.in of your platform, same as before.

Note1:
if you do not specify these switches, the default behaviour of the script
is the same as before, i.e.:
- download the executable in the target
- resume the CPU (if "-debug" is not present)

Note2:
The "simple" command string used before to reset/halt/init the CPU should
work in the same way (if you do not want to use a command file)

Note3:
In the examples for some targets, we do not check the $noload_flag and
$noexec_flag variables, this is because we rely "blindly" on the switches
specified in the command line of the configuration.in (and to not add
clutter to the example code). Anyway, maybe it's more correct to check
them to avoid surprises -- but if you are strictly confident on what
you are doing, they are somewhat a little bit redundant.

Note 4:
All this kind of handling could be avoided by simply write an "openocd.cfg"
file, i.e., perhaps extending that provided by OpenOCD official distribution.
But we want to mantain mainstream files to not touch standard templates,
and anyway a little bit of customization is almost always involved, apart
the simplicty of issuing the "make run" or "make debug" commands in the
shell. Please note that you could still write your openocd.cfg in case,
and still use this script chaining.

New handling for targets listed above will follow in next commits.