Makefile[.fn.in]: remove "nulfile" handling
The infamous "nul" file gets generated in Windos when unproper cmd.exe
commands are issued in an MSYS2 Bash style shell. This problem should now
be gone away since the shell/OS handling is correctly handled.

Delete both the function and the calling part so make operations are a
little bit faster.