Makefile|[.gnatstudio]|sweetada.gpr: a little bit of integration in GNATStdio.
We make Makefile.gnatstudio the default make file in order to have not
listed every target of the master Makefile when you issue Build -> Makefile.

As a reminder, when running GNATStdio in Windos, remember to silence the
OSTYPE Scenario variable setting, or better, OSTYPE = cmd.

In Windos you will need to put a suitable make executable in PATH
(perhaps the one that you could find in libutils/exe/make.zip).

And, of course, you have to configure your platform before running the
IDE (in case you will get errors because configure.gpr is missing).