all qemu platforms/qemu.bat: more GNATStdio integration
When issuing a "make run" in GNATStdio, the IDE spits off repeatedly errors
because the Windos timeout.exe utility does not tolerate being redirected.
So we provide a builtin implementation in the script.

As a random gain, this is a good solution because we are not calling
another external executable. The cons is to waste some CPU cycles,
but it is only for a few secs.