QEMU-AArch64: change the QEMU debug interface
When debugging a virtual QEMU target, the GDB session tooks place in the
same terminal shell from where you are issuing build make commands.

This could be ok, but in some situations this solution is impractical.
Inside GNAT Studio, for example, debugging (by running Build->Makefile->debug)
is impossible because the IDE capture the input/output descriptors of every
process, and segregates the output in a small window view without possibility
of user interaction.

So we spawn a terminal when running GDB.

Unfortunately, as usual, both the environments and IDEs impose every kind
of headaches. Inside the script you can select your preferred terminal
from the choices.

After some test, we will apply those commits also to other QEMU targets,
and maybe we will make the terminal setup a configurable item in order
to have it centralized together with other system attributes.

The qemu.bat has also some variable handling changes, which is better to
not comment, due to absurdities in command line parsing.