share/terminal.sh:
We introduce the terminal.sh, which is a little script to make terminal
handling more suited to adapt various Linux distribution, that may lack
xterm. The user has to peek the preferred terminal to be used in QEMU
serial port and OpenOCD server, and in every other situation where it could
be needed as an independent window carrying a process.

In macs and windos we don't need this since we rely on the standard
Terminal and cmd.exe console respectively.

We will slowly update scripts whenever this will be used.

NOTE: we explicitly export TERMINAL in configuration.in since it is not
exactly a build-system executable (and thus it isn't automatically exported)
but rather a user application, albeit a system-wide one.