share/openocd.tcl
We remove the -w switch, which purpose is to not close the window and,
errors should happen, allows a visual diagnostic. Instead we encode in
the calling script a keypress pause in case of error.

Removing of the eval command which should be always avoided.

Plus, we completely rewrite the calling scripts of every OS, making them
more efficient, simpler and uniform.