share/openocd.py
We remove the -w switch, which purpose is to not close the window and,
errors should happen, allows a visual diagnostic. Instead we encode in
the calling script a keypress pause in case of error.

Plus, we completely rewrite the calling scripts of every OS, making them
more efficient, simpler and uniform.