all platforms with windos-batch-scripts: use terminal.bat to select a terminal console
So far, only ConEmu is selectable, otherwise a standard cmd.exe window will come up.
Enable it in the master configuration.in like the next commit will.

ConEmu (or other consoles) should be visible in path because the script returns only
the executable filename.

Plus, removal of the out-of-context setting of PYTHONHOME environment variable.

Plus, some cosmetics on variable names.