RaspberryPi3/createbootdevice.sh:
We make the createbootdevice.sh script independent from the context, runnable
out of the SweetAda build system.

The script will write an executable (with a JTAG-enable config.txt)
on a SD-card device, either specifiyng its UUID or /dev/???.

It is useful to write the share/rpi3_bootloop.bin stub, which keeps the
CPU hang in a loop. This way you can then download the SweetAda kernel with
normal "make run" or "make debug" operations (after a "make session-start"
to activate the OpenOCD server, of course).