All OpenOCD and QEMU platforms: update scripts
Those scripts try to create an uniform view about the run/debug interface
of SweetAda, to improve the experience. The goal is to make operations
transparent across a CLI and from within an IDE, like GPS.

Care is required to handle running of scripts when launched from an IDE,
which tries to check for all processes spawned and so could show hanging
views or maybe could be stuck in a loop, waiting for processes to complete.

Where possible, the scripts also show some diagnostics when running all the
various components, like OpenOCD servers, GDB executables, and so on.

Furthermore, all terminal windows (apart some QEMU hardwired serial consoles,
windows=PuTTY and osx=terminal) are now under the control of the TERMINAL
variable, so you can choose your preferred terminal emulator, most likely
in configuration.in.

All scripts are nearly identical apart the obvious target-dependent options
for QEMU and GDB.