cpus/AArch64:
We introduce "experimental" preprocessing on this CPU.

Due to the enormous differences and options, it becomes increasingly difficult
to mantain a clean scheme to handle definitions and packages when this kind
of devices come into play. In particular, some registers are not present in
some CPUs, and you cannot easily add/remove them without a ridicoulous amount
of code clutter.

So we start to use GNATPREP on CPU definitions, sectioning out pieces
of text when appropriate.

During the configuration phase, the (un-)preprocessed ad[b|s]p files get
preprocessed with the aid of a CPU_DEF file, stated in configuration.

So far, only the Cortex-A53 is listed. For another CPU, we only have to
write another CPU_DEF file and enable it with if/ifeq in the CPU's
configuration.in, which in turn inherits parameters from the platform's
configuration.in -- where you state which CPU you want.

Note that the unpreprocessed "p" files, and the CPU_DEF, are part of the
configuration files, so if they get modified, a whole configuration and
compilation from scratch will be triggered. This is more than appropriate,
because the CPU files could depend on configuration directives that may
affect the whole CPU setup, compilation switches, and so on, avoiding the
risk of object files already compiled that could be out of sync.

By the way, we define some registers that describe the features implemented
in the ARMv8A target. Example of this handling in the BSP of QEMU-AArch64
in the next commit.