Makefile:
We declare the entry point label of the executable as a variable. Its value
is by default the standard "_start" symbol, but now you can declare it
whatever you want, just override KERNEL_ENTRY_POINT in the configuration file.

This will allow us to remove the hardwired symbol name from various script
(e.g., the OpenOCD executor and others). Anyway, it's useful to have it
configurable. The entry point symbol will be printed out in the cfg dump,
and its value will be printed out at the end of the compilation phase.

We also export TOOLCHAIN_NM so that it can be used by the elftool utility.