Veecom: new platform
Maybe you want to see electrons on the wire driven by Ada code, so here
it is a quick-and-dirty port to Veecom, a RISC-V processor simulated in
the Logisim-evolution digital simulator.

https://github.com/logisim-evolution/logisim-evolution
https://github.com/RiscWiz5/Veecom

Just load the kernel.rom binary image in the memory block and start the
simulation.

Obviously the simulation is somewhat slow -- since is working down to the
logical ports -- so wait a minute before seeing characters on the console.