Makefile: add two new variables to the list of exports
GPR_CORE_CPU is already defined, inherited from <cpu>/configuration.in.
We make it a full citizen, exporting it in the Master Makefile rather
than export it every time in the various CPU configuration files. We will
slowly remove the explicit export since it doesn't hurt.

GCCDEFINES_CPU is a new variable, and will be defined as well in
<cpu>/configuration.in when needed. This way we can configure the list
of GCC defines to be created in the core gcc-defines.ads on a CPU-basis,
since it is a little bit exaggerated to list every CPU-dependent macro
of all CPUs, that is going to be surely ignored in CPU-agnostic code.
Of course the macros for major CPU classes assertion are kept by default.
Example in the next commit dealing with M68k.